Pass email to security form on reset in frmResetAccount
The frmResetAccount form now passes the email entered by the user to the frmResetAccountSecurity form when the reset button is clicked. A new public method, SetEmail, was added to the frmResetAccountSecurity class to set the text of the lblWelcome1 label to the provided email, allowing the email to be displayed on the form.